Delete redundant file, change formatting, move Contests Winners under Contest, added links to doc
