+++ App/HejDTO/Sources/HejDTO/Contests/GET/ContestWinner/ContestWinners.swift
@@ -1,18 +1,26 @@
 
 import Foundation
 
-public struct ContestWinners: Codable, Equatable {
+extension Contests {
     
-    public let page: Int
-    public let limit: Int
-    public let pages: Int
-    public let total: Int
-    public let links: Common.Links
-    public let embedded: ContestWinners.Embedded
-    
-    private enum CodingKeys: String, CodingKey {
-        case page, limit, pages, total
-        case links = "_links"
-        case embedded = "_embedded"
+    ///https://docs.hejto.pl/#tag/Contests/operation/get_contest_winners
+    public struct ContestWinners: Codable, Equatable {
+        
+        public struct Embedded: Codable, Equatable {
+            public let items: [ContestWinner]
+        }
+        
+        public let page: Int
+        public let limit: Int
+        public let pages: Int
+        public let total: Int
+        public let links: Common.Links
+        public let embedded: Contests.ContestWinners.Embedded
+        
+        private enum CodingKeys: String, CodingKey {
+            case page, limit, pages, total
+            case links = "_links"
+            case embedded = "_embedded"
+        }
     }
 }
