--- App/HejDTO/Sources/HejDTO/Contests/GET/Contests/Contests.Links.swift
@@ -1,15 +1,0 @@
-
-import Foundation
-
-extension Contests {
-    
-    public struct Links: Codable, Equatable {
-        
-        public let selfLink: Common.Link
-        public let follows: Common.Link
-        
-        private enum CodingKeys: String, CodingKey {
-            case selfLink = "self", follows
-        }
-    }
-}
