+++ App/HejDTO/Tests/HejDTOTests/Contests/GET/ContestsDetails.Tests.swift
@@ -116,7 +116,7 @@ final class ContestsDetailsTests: XCTestCase {
                 verified: true,
                 sponsor: true,
                 createdAt: .date2019_08_24T141522Z,
-                links: Contests.Links(selfLink: Common.Link(href: .string),
+                links: Contests.ContestOrganizer.Links(selfLink: Common.Link(href: .string),
                                       follows: Common.Link(href: .string))),
             userScore: 0,
             userPosition: 0,
